tests/requirements.txt
google-genai
pillow
pytest
python-dotenv
pytest-asyncio
